more styling, fixed the playlist query being out of sync with the new database view changes

+++ core/src/main/java/com/ouchadam/fang/domain/ItemToPlaylist.java
@@ -4,19 +4,19 @@ import com.ouchadam.fang.domain.item.Item;
 
 public class ItemToPlaylist {
 
-    private final int itemId;
+    private final long itemId;
     private final long downloadId;
 
     public static ItemToPlaylist from(Item item, long downloadId) {
         return new ItemToPlaylist(item.getId(), downloadId);
     }
 
-    public ItemToPlaylist(int itemId, long downloadId) {
+    public ItemToPlaylist(long itemId, long downloadId) {
         this.itemId = itemId;
         this.downloadId = downloadId;
     }
 
-    public int getItemId() {
+    public long getItemId() {
         return itemId;
     }
